Style fix. (+12 squashed commits) Squashed commits: [d8cc1ee] Forgot to add json_encode to return a string. [5ef6fce] Fixed docs. [1ac510e] Style fix. [a5fba3a] Removed parent construct from broadcasters. [b7047c8] Updates. [22900c1] Style fixes. [ddf4ba4] Added an abstract class for Broadcasters. [1c23e53] Updated Broadcast Manager to cache a socket_id as "broadcast:socket..." instead of "pusher:socket..." [180934c] Updated the validAuthenticationResponse method to send back data as "channel_data" [0712630] Update RedisBroadcaster.php
Added support for `$this->dontBroadcastToCurrentUser();`

With the addition of the socket_id in the payload, socket.io can broadcast the event as the requesting user.
[6e43303] Code style updates
[237cacd] Adding support for new Broadcast Support Provider
