+++ src/Illuminate/Broadcasting/BroadcastManager.php
@@ -86,7 +86,7 @@ class BroadcastManager implements FactoryContract
         }
 
         return $this->app['cache']->get(
-            'pusher:socket:'.$request->session()->getId()
+            'broadcast:socket:'.$request->session()->getId()
         );
     }
 
@@ -105,7 +105,7 @@ class BroadcastManager implements FactoryContract
         $request = $request ?: $this->app['request'];
 
         $this->app['cache']->forever(
-            'pusher:socket:'.$request->session()->getId(), $request->socket_id
+            'broadcast:socket:'.$request->session()->getId(), $request->socket_id
         );
     }
