+++ src/Illuminate/Broadcasting/Broadcasters/Broadcaster.php
@@ -1,0 +1,101 @@
+<?php
+
+namespace Illuminate\Broadcasting\Broadcasters;
+
+use Illuminate\Support\Str;
+use Illuminate\Contracts\Broadcasting\Broadcaster;
+use Symfony\Component\HttpKernel\Exception\HttpException;
+
+abstract class Broadcaster implements Broadcaster
+{
+    /**
+     * The registered channel authenticators.
+     *
+     * @var array
+     */
+    protected $channels = [];
+
+    /**
+     * Register a channel authenticator.
+     *
+     * @param  string  $channel
+     * @param  callable  $callback
+     * @return $this
+     */
+    public function auth($channel, callable $callback)
+    {
+        $this->channels[$channel] = $callback;
+
+        return $this;
+    }
+
+    /**
+     * Authenticate the incoming request for a given channel.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return mixed
+     */
+    public function check($request)
+    {
+        $channel = str_replace(['private-', 'presence-'], '', $request->channel_name);
+
+        foreach ($this->channels as $pattern => $callback) {
+            if (! Str::is($pattern, $channel)) {
+                continue;
+            }
+
+            $parameters = $this->extractAuthParameters($pattern, $channel);
+
+            if ($result = $callback($request->user(), ...$parameters)) {
+                return $this->validAuthenticationResponse($request, $result);
+            }
+        }
+
+        throw new HttpException(403);
+    }
+
+    /**
+     * Return the valid Pusher authentication response.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  mixed  $result
+     * @return mixed
+     */
+    protected function validAuthenticationResponse($request, $result)
+    {
+        if (is_bool($result)) {
+            return json_encode($result);
+        }
+
+        return json_encode(['channel_data' => $result]);
+    }
+
+    /**
+     * Extract the parameters from the given pattern and channel.
+     *
+     * @param  string  $pattern
+     * @param  string  $channel
+     * @return array
+     */
+    protected function extractAuthParameters($pattern, $channel)
+    {
+        if (! Str::contains($pattern, '*')) {
+            return [];
+        }
+
+        $pattern = str_replace('\*', '([^\.]+)', preg_quote($pattern));
+
+        if (preg_match('/^'.$pattern.'/', $channel, $keys)) {
+            array_shift($keys);
+
+            return $keys;
+        }
+
+        return [];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    abstract public function broadcast(array $channels, $event, array $payload = []);
+}
