+++ src/Illuminate/Broadcasting/Broadcasters/PusherBroadcaster.php
@@ -5,10 +5,8 @@ namespace Illuminate\Broadcasting\Broadcasters;
 use Pusher;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
-use Illuminate\Contracts\Broadcasting\Broadcaster;
-use Symfony\Component\HttpKernel\Exception\HttpException;
 
-class PusherBroadcaster implements Broadcaster
+class PusherBroadcaster extends Broadcaster
 {
     /**
      * The Pusher SDK instance.
@@ -18,60 +16,13 @@ class PusherBroadcaster implements Broadcaster
     protected $pusher;
 
     /**
-     * The registered channel authenticators.
-     *
-     * @var array
-     */
-    protected $channels = [];
-
-    /**
      * Create a new broadcaster instance.
-     *
      * @param  \Pusher  $pusher
      * @return void
      */
     public function __construct(Pusher $pusher)
     {
         $this->pusher = $pusher;
-    }
-
-    /**
-     * Register a channel authenticator.
-     *
-     * @param  string  $channel
-     * @param  callable  $callback
-     * @return $this
-     */
-    public function auth($channel, callable $callback)
-    {
-        $this->channels[$channel] = $callback;
-
-        return $this;
-    }
-
-    /**
-     * Authenticate the incoming request for a given channel.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return mixed
-     */
-    public function check($request)
-    {
-        $channel = str_replace(['private-', 'presence-'], '', $request->channel_name);
-
-        foreach ($this->channels as $pattern => $callback) {
-            if (! Str::is($pattern, $channel)) {
-                continue;
-            }
-
-            $parameters = $this->extractAuthParameters($pattern, $channel);
-
-            if ($result = $callback($request->user(), ...$parameters)) {
-                return $this->validAuthenticationResponse($request, $result);
-            }
-        }
-
-        throw new HttpException(403);
     }
 
     /**
@@ -90,30 +41,6 @@ class PusherBroadcaster implements Broadcaster
                 $request->channel_name, $request->socket_id, $request->user()->id, $result
             );
         }
-    }
-
-    /**
-     * Extract the parameters from the given pattern and channel.
-     *
-     * @param  string  $pattern
-     * @param  string  $channel
-     * @return array
-     */
-    protected function extractAuthParameters($pattern, $channel)
-    {
-        if (! Str::contains($pattern, '*')) {
-            return [];
-        }
-
-        $pattern = str_replace('\*', '([^\.]+)', preg_quote($pattern));
-
-        if (preg_match('/^'.$pattern.'/', $channel, $keys)) {
-            array_shift($keys);
-
-            return $keys;
-        }
-
-        return [];
     }
 
     /**
